Fix bugs
Firstly, the playScore and computerScore variables were not defined
as arguments, since they are not defined within the function.
the playRound() function.

Secondly, the playerSelection and computerSelection variables needed
to be defined within the while loop in the game() function,

Thirdly, the round variable was incrementing outside of the
while loop.

Lastly, the if statement that checks for the winner was outside
of the game() function when it needed to be inside.